requirements.txt
pyyaml
pyserial
requests
click
